Made Requested changes + More misc cleanup
